CM-107: Voting: split actions and thunks
