+++ packages/core/redux/watcher/actions.js
@@ -10,7 +10,7 @@ import { initMainWallet } from '../mainWallet/actions'
 import { watchInitMarket } from '../market/actions'
 import { initTokens } from '../tokens/actions'
 import { initDAOs } from '../daos/actions'
-import { watchInitPolls } from '../voting/actions'
+import { watchInitPolls } from '../voting/thunks'
 import { initMultisigWalletManager } from '../multisigWallet/actions'
 import { initWallets } from '../wallets/actions'
 import { WATCHER } from './constants'
