+++ src/components/dialogs/PublishPollDialog/PublishPollDialog.jsx
@@ -10,7 +10,7 @@ import { Translate } from 'react-redux-i18n'
 import { connect } from 'react-redux'
 import { modalsClose } from 'redux/modals/actions'
 import { PTPoll } from '@chronobank/core/redux/voting/types'
-import { activatePoll } from '@chronobank/core/redux/voting/actions'
+import { activatePoll } from '@chronobank/core/redux/voting/thunks'
 import ModalDialog from '../ModalDialog'
 import './PublishPollDialog.scss'
 import { prefix } from './lang'
