+++ src/components/voting/EndPollTopButton/EndPollTopButton.jsx
@@ -8,7 +8,7 @@ import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { getSelectedPoll } from '@chronobank/core/redux/voting/selectors'
 import { Translate } from 'react-redux-i18n'
-import { endPoll } from '@chronobank/core/redux/voting/actions'
+import { endPoll } from '@chronobank/core/redux/voting/thunks'
 import { isCBE } from '@chronobank/core/redux/session/selectors'
 import { PTPoll } from '@chronobank/core/redux/voting/types'
 import Button from 'components/common/ui/Button/Button'
