+++ src/components/voting/Poll/Poll.jsx
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
-import { activatePoll, endPoll, removePoll, selectPoll } from '@chronobank/core/redux/voting/actions'
+import { activatePoll, endPoll, removePoll, selectPoll } from '@chronobank/core/redux/voting/thunks'
 import { PTPoll } from '@chronobank/core/redux/voting/types'
 import BigNumber from 'bignumber.js'
 import { prefix } from './lang'
