+++ src/components/voting/PollEditForm/PollEditForm.jsx
@@ -24,7 +24,7 @@ import { daoByType } from '@chronobank/core/redux/daos/selectors'
 import VotingManagerDAO from '@chronobank/core/dao/VotingManagerDAO'
 import { TX_CREATE_POLL } from '@chronobank/core/dao/constants/VotingManagerDAO'
 import { DUCK_SESSION } from '@chronobank/core/redux/session/constants'
-import { createPoll } from '@chronobank/core/redux/voting/actions'
+import { createPoll } from '@chronobank/core/redux/voting/thunks'
 import { DUCK_VOTING } from '@chronobank/core/redux/voting/constants'
 import { modalsClose } from 'redux/modals/actions'
 import Amount from '@chronobank/core/models/Amount'
