+++ src/layouts/partials/VoteHistoryContent/VoteHistoryContent.jsx
@@ -9,7 +9,7 @@ import Button from 'components/common/ui/Button/Button'
 import PropTypes from 'prop-types'
 import { getSelectedPoll, getVotingFlags } from '@chronobank/core/redux/voting/selectors'
 import { initAssetsHolder } from '@chronobank/core/redux/assetsHolder/actions'
-import { goToVoting, listPolls, vote } from '@chronobank/core/redux/voting/actions'
+import { goToVoting, listPolls, vote } from '@chronobank/core/redux/voting/thunks'
 import { isCBE } from '@chronobank/core/redux/session/selectors'
 import { getDepositAmount } from '@chronobank/core/redux/assetsHolder/selectors'
 import { PTPoll } from '@chronobank/core/redux/voting/types'
