+++ src/layouts/partials/VotingContent/VotingContent.jsx
@@ -12,7 +12,7 @@ import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
 import { getPollsDetailsList, getVotingFlags } from '@chronobank/core/redux/voting/selectors'
 import { initAssetsHolder } from '@chronobank/core/redux/assetsHolder/actions'
-import { listPolls } from '@chronobank/core/redux/voting/actions'
+import { listPolls } from '@chronobank/core/redux/voting/thunks'
 import { getAccount, isCBE } from '@chronobank/core/redux/session/selectors'
 import { getDepositAmount } from '@chronobank/core/redux/assetsHolder/selectors'
 import { PTPoll } from '@chronobank/core/redux/voting/types'
